password-generator/python/requirements.txt
greenlet
pandas
psycopg2
python-dateutil
python-dotenv
pytz
six
sqlalchemy==1.4.22
typing_extensions
tzdata
bcrypt==4.2.0
cffi==1.17.0
cryptography==43.0.0
greenlet==3.0.3
numpy==2.0.1
pandas==2.2.2
paramiko==3.4.0
psycopg2==2.9.9
pycparser==2.22
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
tzdata==2024.1
